Prohibit ArrowFunc nested in exprs except CondExpr
Thus, move it out of PrimaryExpression, to the top-level Expression
rule. This matches JS, which puts ArrowFunction at the level of
assignment operators: https://tc39.es/ecma262/#sec-assignment-operators

This actually makes more sense for us than for JS, because JS at least
has short-circuiting logical operators where it could potentially make
sense to do `flag || x => 2*x`, whereas in Mechanical that's a type
error (logical operators may only take boolean values). (In fact, in JS
it isn't even a type error to do `-(x => 2*x)`, you just get `NaN`,
although it's not very useful.)

A potential, theoretical benefit of this restriction is it might put us
in a position to be able to expand arrow-function syntax with prefix
symbols, if for some reason we want to do that. I mean, in general it's
just a good idea to lean towards making the syntax more restrictive than
not, since it's easier to add than remove valid syntax.