Update dependencies: mocha (security), ts-node (perf)
Security advisory from GitHub for minimist, a dependency of Mocha:
  https://github.com/laughinghan/mechanical/network/alert/package-lock.json/minimist/open
  https://github.com/mochajs/mocha/releases/tag/v7.1.1

ts-node improved performance in v8.7 and v8.8:
  https://github.com/TypeStrong/ts-node/issues/754#issuecomment-602119414

As detailed in the comment I left, unfortunately it was a limited
improvement for me since my tests are so fast already. For now I'm doing
TS_NODE_TRANSPILE_ONLY=true to run the tests faster without bothering
with typechecking, and relying on VS Code's TypeScript integration for
typechecking. Not sure if I should build that into the package.json
`npm test` script though.